refactor(neon): share credential and profile code across every CLI, without a package
`neon`, `@neon/env` and `neon-init` each had their own answer to "read the stored
credential", which is how they came to disagree: `neon-env` read `DEFAULT` only and had no
`--profile`, so `neon --profile dbx env` and `neon-env` could resolve different accounts on
the same machine, and `neon-init` looked only for `access_token`, so an account signed in
with an API key read as not signed in and got sent to a browser.

`shared/cli-core/src` is now the one implementation, and it is deliberately **not a
package**. `scripts/sync-shared.mjs` copies it into each consumer's `src/_shared` before
they compile, that copy is gitignored, and the imports are relative — so the code is built
into every `dist` and nothing resolves at runtime.

A workspace package cannot do this here, and it is worth recording why so nobody retries it:
every package builds with `bundle: false` or plain `tsc`, so a bare specifier survives into
`dist` and has to resolve from `node_modules`, which an unpublished package cannot do for
someone who installed `neon` from npm. `bundledDependencies` exists for exactly that and
pnpm refuses it — `ERR_PNPM_BUNDLED_DEPENDENCIES_WITHOUT_HOISTED` — needing a workspace-wide
`nodeLinker: hoisted`. Publishing it would put an internal surface on the registry, and
`@neon/config`, the one published package this could hang off, is consumer-facing. Verified
all three of those rather than assumed them.

What moved: `credentials.ts`, `profiles.ts`, `auth_selection.ts`, `secure_file.ts`, and the
config-path resolution that was in `@neon/config/src/paths.ts`. `@neon/config/paths`
re-exports the path half **explicitly rather than with `export *`**, so its public API is
byte-identical and the credential paths and ownership checks the CLIs need do not leak into
it — checked by comparing the built module's exports before and after.

What it buys:

- `neon-env` takes `--profile` and honours `NEON_PROFILE`, resolving the same credential the
  `neon` CLI would. Four tests cover the precedence.
- `neon-init` reads through the shared reader, so it understands both credential kinds and
  no longer carries its own copy of the path resolution — the copy AGENTS.md described as
  temporary.
- The profile and precedence rules exist once, so the next change to them cannot land in one
  CLI and miss the others.

`shared/cli-core/README.md` carries the rules, and AGENTS.md points at it.
