+++ .gitignore
@@ -5,3 +5,6 @@ coverage
 .env.*
 !.env.example
 CLAUDE.md
+
+# Generated by scripts/sync-shared.mjs from shared/cli-core — never edit or commit.
+packages/*/src/_shared/
