+++ AGENTS.md
@@ -298,11 +298,31 @@ the same thing with a subpath:
 directory" and all three disagreed: the CLI honoured `XDG_CONFIG_HOME` but not
 `NEONCTL_CONFIG_DIR`, `@neon/env` honoured the env var but not XDG, and `neon-init`
 hardcoded `~/.config/neonctl`. With `XDG_CONFIG_HOME` set, the CLI wrote credentials
-somewhere the other two never looked. Add a reader to that module rather than to a fourth
-private copy. (`neon-init` still has an inline copy, deliberately — it has no workspace
-dependencies and taking one on `@neon/config` would pull `@neon/sdk`, `zod` and `jiti` into
-its install footprint for a file path. That copy disappears when the package folds into
-`packages/cli`.)
+somewhere the other two never looked. **That implementation now lives in `shared/cli-core`
+and `@neon/config/paths` re-exports it** — see below.
+
+### `shared/cli-core` — code every CLI compiles as its own
+
+Credential reading, profile resolution and config paths are shared by `neon`, `@neon/env`,
+`neon-init` and `@neon/config` from `shared/cli-core/src`. It is **not a package**:
+`scripts/sync-shared.mjs` copies it into each consumer's `src/_shared` before they build,
+that copy is gitignored, and the imports are relative, so the code is compiled into every
+`dist` and nothing resolves at runtime.
+
+It is not a workspace package because it cannot be one. Every package here builds with
+`bundle: false` or plain `tsc`, so a bare specifier survives into `dist` and must resolve
+from `node_modules` — which an unpublished package cannot do for anyone who installed `neon`
+from npm. `bundledDependencies` is the mechanism for exactly that and pnpm refuses it here
+(`ERR_PNPM_BUNDLED_DEPENDENCIES_WITHOUT_HOISTED`; it needs `nodeLinker: hoisted`). Publishing
+it would put an internal surface on the registry, and `@neon/config` — the one published
+package all of this could hang off — is consumer-facing.
+
+**Edit `shared/cli-core/src`, never `packages/*/src/_shared`.** Keep it dependency-free (Node
+builtins only, because `neon-init` has no workspace dependencies), and keep loggers, yargs and
+API clients out of it — take a callback or a value instead. Its tests live in `packages/cli`,
+the one consumer that runs them. `@neon/config/paths` re-exports it **explicitly rather than
+with `export *`**, so the credential paths and ownership checks the CLIs need do not become
+public API the next time something is added.
 
 The directory is `neon`; `neonctl` is the pre-rename name and is **read forever, in place**.
 Nothing is moved, copied, or deleted, so a second copy of a credential can never go stale
