+++ shared/cli-core/src/auth_selection.ts
