+++ shared/cli-core/src/secure_file.ts
